Refactor handling of URL search params into a custom hook
